Pasting from a large comment:
The following line was commented out by Scott Wheeler <wheeler@kde.org>
We don't actually need to store all of the pages / packets in memory since
(a) libtheora doesn't use them anyway in the one call that we make after this
that usese t_state and (b) it basically buffers the entire file to memory if
we queue them up like this and that sucks where a typical file size is a few
hundred megs.

Also this thing probably wouldn't be so darn slow if it started searching for
the end of the stream at the end of the file, rather than stepping through
the entire file, but I don't know if libogg can do that.  If not I may be
forced to get wacky with the Ogg paging implementation in TagLib.  ;-)

CCMAIL:bj@altern.org

svn path=/trunk/kdemultimedia/kfile-plugins/; revision=356342