Some TODOs
Since a m3u file is just a textfile, make sure to check the file extension.
But doing this results in a segmentation fault at the moment.
Well, when I return in the startAnalysis function, I get the segfault.

I'm not sure why this happens, so comment the code and add a todo.


svn path=/trunk/KDE/kdemultimedia/kfile-plugins/; revision=646366